it's int

+++ src/main/java/ink/ikx/rt/impl/events/SanityChangeEvent.java
@@ -17,7 +17,7 @@ public class SanityChangeEvent extends Event {
         this.originalSanity = originalSanity;
     }
 
-    public float getOriginalSanity() {
+    public int getOriginalSanity() {
         return originalSanity;
     }
